+++ GDO.Apps.Maps/Core/Sources/Images/StaticImageSource.cs
@@ -10,8 +10,8 @@ namespace GDO.Apps.Maps.Core.Sources.Images
         public int Width { get; set; }
         public int Height { get; set; }
         public string URL { get; set; }
-        public float[] Extent { get; set; }
-        new public void Modify(int width, int height, string url, float[] extent)
+        public double[] Extent { get; set; }
+        new public void Modify(int width, int height, string url, double[] extent)
         {
             Width = width;
             Height = height;
